+++ src/ebay-lightbox-dialog/lightbox-dialog.tsx
@@ -1,6 +1,11 @@
-import React, { CSSProperties, FC } from 'react'
+import React, {
+    CSSProperties,
+    FC
+} from 'react'
 import classNames from 'classnames'
 import { DialogBaseProps, DialogBaseWithState } from '../ebay-dialog-base'
+import { DialogCloseEventHandler } from '../ebay-dialog-base/types'
+
 
 type Mode = 'default' | 'mini'
 type Size = 'wide' | 'narrow' | 'fullscreen' | 'large'
@@ -11,7 +16,7 @@ export type Props<T = any> = Omit<DialogBaseProps<T>, 'size'> & {
   size?: Size;
   bannerImgSrc?: string;
   bannerImgPosition?: CSSProperties['backgroundPosition'];
-  onClose?: () => void;
+  onClose?: DialogCloseEventHandler;
 }
 
 const EbayLightboxDialog: FC<Props> = ({